double quotes

+++ .eslintrc.js
@@ -13,7 +13,7 @@ module.exports = {
     "no-debugger": process.env.NODE_ENV === "production" ? "error" : "off",
     "prettier/prettier": "error",
     semi: ["error", "never"],
-    quotes: ["error", "single"]
+    quotes: ["error", "double"]
   },
   parserOptions: {
     parser: "babel-eslint"
